# V2-node/API/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node Goku-e-commerce API",
    "description": "API for user registration, login, and address management",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/users": {
      "post": {
        "summary": "Create a new user",
        "description": "Creates a new user with the given details.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "description": "User details for registration",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "name": {
                    "type": "string",
                    "example": "User"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User created"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error creating user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error creating user"
                    },
                    "error": {
                      "type": "string",
                      "example": "Email already used"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all users",
        "description": "Retrieves a list of all registered users.",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Users retrieved"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "email": {
                            "type": "string",
                            "example": "user@example.com"
                          },
                          "name": {
                            "type": "string",
                            "example": "User"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error retrieving users"
                    },
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing user",
        "description": "Updates the details of an existing user by ID.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to update",
            "schema": {
              "type": "string",
              "example": "66be623653908d88151ee29b"
            }
          }
        ],
        "requestBody": {
          "description": "User details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "newemail@example.com"
                  },
                  "name": {
                    "type": "string",
                    "example": "New Name"
                  },
                  "password": {
                    "type": "string",
                    "example": "newpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User updated"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "newemail@example.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "New Name"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Failed to update user"
                    },
                    "error": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "description": "Deletes an existing user by ID.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user to delete",
            "schema": {
              "type": "string",
              "example": "66be623653908d88151ee29b"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    },
                    "error": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/address": {
      "post": {
        "summary": "Add an address to a user",
        "description": "Adds a new address to a user based on the provided data.",
        "tags": [
          "Addresses "
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Address data to add",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cep": {
                    "type": "string",
                    "example": "06454080"
                  },
                  "id": {
                    "type": "string",
                    "example": "userId123"
                  }
                },
                "required": [
                  "cep",
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address added"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        },
                        "addresses": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "zip": {
                                "type": "string",
                                "example": "06454080"
                              },
                              "street": {
                                "type": "string",
                                "example": "Alameda Itapecuru"
                              },
                              "city": {
                                "type": "string",
                                "example": "Barueri"
                              },
                              "state": {
                                "type": "string",
                                "example": "SP"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error adding address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error adding address"
                    },
                    "error": {
                      "type": "string",
                      "example": "Invalid CEP"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove an address from a user",
        "description": "Removes an address from a user based on the provided IDs.",
        "tags": [
          "Addresses "
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "User and address IDs to remove the address",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "userId123"
                  },
                  "addressId": {
                    "type": "string",
                    "example": "addressId123"
                  }
                },
                "required": [
                  "userId",
                  "addressId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address removed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid request parameters"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/address/{cep}": {
      "get": {
        "summary": "Get address information by CEP",
        "description": "Retrieves address information based on the provided CEP.",
        "tags": [
          "Addresses "
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cep",
            "in": "path",
            "required": true,
            "description": "CEP to retrieve address information",
            "schema": {
              "type": "string",
              "example": "06454080"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "zip": {
                      "type": "string",
                      "example": "06454080"
                    },
                    "street": {
                      "type": "string",
                      "example": "Rua Teste"
                    },
                    "city": {
                      "type": "string",
                      "example": "Cidade Teste"
                    },
                    "state": {
                      "type": "string",
                      "example": "TS"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Address not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Address not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate user",
        "description": "Logs in a user and returns a JWT token if credentials are valid.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "description": "User credentials for login",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful and token returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "jwt-token-here"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}